requirements.txt
pyperclip==1.8.0
pycryptodome==3.9.8
